requirements.txt
# Deep Learning
ultralytics==8.3.178
torch==2.9.0.dev20250812
torchaudio==2.8.0.dev20250812
torchvision==0.24.0.dev20250812
--extra-index-url https://download.pytorch.org/whl/nightly/cpu
onnx>=1.12.0,<1.18.0
onnxslim >=0.1.59
onnxruntime==1.22.1
coremltools>=8.0
lap>=0.5.12

# Model Tracking
mlflow==3.2.0

# Log
loguru==0.7.3

# Env
python-dotenv==1.1.1

pytube==15.0.0